Fix PydanticUndefinedAnnotation
