+++ app/api/notes/schema.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from pydantic import BaseModel, Field
 
 from app.models import NoteSchema
