+++ app/models/__init__.py
@@ -1,3 +1,4 @@
 from .base import Base
-from .notebooks import Notebook, NotebookSchema
-from .notes import Note, NoteSchema
+from .notebooks import Notebook
+from .notes import Note
+from .schema import NotebookSchema, NoteSchema
